requirements.txt
aiogram==2.22.2
aiohttp==3.8.4
aiosignal==1.3.1
altgraph==0.17.3
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==3.1.0
et-xmlfile==1.1.0
frozenlist==1.3.3
idna==3.4
loguru==0.7.0
macholib==1.16.2
magic-filter==1.0.9
multidict==6.0.4
openpyxl==3.1.2
pyinstaller==5.9.0
pyinstaller-hooks-contrib==2023.2
pytz==2023.3
PyYAML==6.0
yarl==1.8.2
python-dotenv==1.0.0
aioschedule==0.5.2